Set up bigtable IT's to run only when IT's STORE_TYPE=BIGTABLE

+++ .utility/gcloud-init.sh
@@ -1,27 +1,33 @@
 #!/bin/bash
-# Download the google cloud SDK
-wget https://dl.google.com/dl/cloudsdk/channels/rapid/downloads/google-cloud-sdk-135.0.0-linux-x86_64.tar.gz
-gunzip google-cloud-sdk-135.0.0-linux-x86_64.tar.gz
-tar xvf google-cloud-sdk-135.0.0-linux-x86_64.tar
-rm google-cloud-sdk-135.0.0-linux-x86_64.tar
+set -ev
 
-# install the beta component & emulator
-cat <<EOF | ./google-cloud-sdk/bin/gcloud components install beta
+if [ "$STORE_TYPE" == "BIGTABLE" ]; then
+	if [ ! -d ./google-cloud-sdk ]; then
+		# Download the google cloud SDK
+		wget https://dl.google.com/dl/cloudsdk/channels/rapid/downloads/google-cloud-sdk-135.0.0-linux-x86_64.tar.gz
+		gunzip google-cloud-sdk-135.0.0-linux-x86_64.tar.gz
+		tar xvf google-cloud-sdk-135.0.0-linux-x86_64.tar
+		rm google-cloud-sdk-135.0.0-linux-x86_64.tar
+	fi
+	
+	# install the beta component & emulator
+	cat <<EOF | ./google-cloud-sdk/bin/gcloud components install beta
 Y
 EOF
 
-cat <<EOF | ./google-cloud-sdk/bin/gcloud beta emulators bigtable env-init --quiet
+	cat <<EOF | ./google-cloud-sdk/bin/gcloud beta emulators bigtable env-init --quiet
 Y
 EOF
 
-# start the emulator
-./google-cloud-sdk/bin/gcloud beta emulators bigtable start &
+	# start the emulator
+	./google-cloud-sdk/bin/gcloud beta emulators bigtable start &
 
-# wait a few seconds
-sleep 10
+	# wait a few seconds
+	sleep 10
 
-# get the emulator port and set it in the env
-./google-cloud-sdk/bin/gcloud beta emulators bigtable env-init > exportBigtableEnv
+	# get the emulator port and set it in the env
+	./google-cloud-sdk/bin/gcloud beta emulators bigtable env-init > exportBigtableEnv
 
-# this next step has to be run outside this script:
-# source exportBigtableEnv
+	# this next step has to be run outside this script:
+	# source exportBigtableEnv
+fi
